% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Match two data bases based on names and dates of birth permitting differences}
\usage{
fuzzy_match(
  query,
  target = NULL,
  cutoff = 0.2,
  total.max = 8,
  full.max = 8,
  check.truncated = TRUE,
  truncate = "am",
  min.n = 25
)
}
\arguments{
\item{query}{A data.table with the names to be matched}

\item{target}{A data.table with the names to be matched to}

\item{cutoff}{Remove any match with score below this number}

\item{total.max}{Maximum Number of erros to permit to be considered}

\item{full.max}{Maximum number of errors permited by full match}

\item{check.truncated}{Checks if second last name is truncated. If truncated matches only remaining part.}

\item{truncate}{The columns to truncate. Defaults to second name `am`}

\item{min.n}{Minimum number of data points needed to fit calibration model}
}
\value{
A data.table with matches
}
\description{
Match two data bases based on names and dates of birth permitting differences
}
