% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_match.R
\name{perfect_match}
\alias{perfect_match}
\title{Match two data bases based on names and dates of birth}
\usage{
perfect_match(query, target = NULL)
}
\arguments{
\item{query}{A data.table with the names to be matched}

\item{target}{A data.table with the names to be matched to}
}
\value{
A data.table with matches
}
\description{
Match two data bases based on names and dates of birth
}
