% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_table.R
\name{wrangle_table}
\alias{wrangle_table}
\title{Clean up names in preparation for match}
\usage{
wrangle_table(tab)
}
\arguments{
\item{tab}{a data.table with the names}
}
\description{
Clean up names in preparation for match
}
